config: simplify route preference parsing
Utilize the extracted implementation of the `parse_ra_route_preference`
function also in the interface configuration parser, where the code was
originally located.

The data type of the route preference has been changed from `int` to
a smaller `int8_t`. Since only three values are used (-1, 0, 1) and
only 3 bits are available to store them in the RA message, an 8-bit
data type is sufficient.

+++ src/odhcpd.h
@@ -439,7 +439,7 @@ struct interface {
 	uint8_t pio_filter_length;
 	struct in6_addr pio_filter_addr;
 	int default_router;
-	int route_preference;
+	int8_t route_preference;
 	uint32_t ra_maxinterval;
 	uint32_t ra_mininterval;
 	uint32_t ra_lifetime;
